feat(launch): generic SAC_GIT_* -> GIT_* env alias in the container shell wrapper
The shell wrapper built by build_inner_argv() now ALWAYS runs a fixed,
values-agnostic alias step first: if SAC_GIT_AUTHOR_NAME (and the other 4
SAC_GIT_* vars) are present in the shell env at that point, it also exports
them under git's own literal names (GIT_AUTHOR_NAME, GIT_AUTHOR_EMAIL,
GIT_COMMITTER_NAME, GIT_COMMITTER_EMAIL, GIT_SSH_COMMAND) -- because git
itself only reads the unprefixed names, while the operator wants the
canonical/introspectable source-of-truth vars to carry a SAC_ prefix
(`printenv | grep SAC_` shows every sac-relevant var in one shot).

This step does not know or care where SAC_GIT_* came from: a project's own
.envrc (direnv source_up, e.g. to ~/proj/.envrc) is the actual current
source and stays entirely outside this module's concern/untouched by this
PR. ${VAR:-} + -n guards make an unset SAC_GIT_* a silent no-op.

Since the alias must apply to EVERY agent (not just ones with
startup_commands), build_inner_argv's bash -lc wrap is now unconditional
instead of early-returning the bare tini/runner argv when
startup_commands is empty -- the alias steps run before any `set -e`
_format_shell_steps may emit, so a false `[ -n ... ]` never aborts launch.

Split the interactive-TUI argv/channel-plan helpers out of
_apptainer_inner_argv.py into a new _apptainer_inner_argv_tui.py (pure
extraction, no behavior change) to stay under the repo's 512-line cap
after adding the alias step; every existing import path is re-exported
unchanged.

Updated the affected test suites for the now-unconditional wrap (the
runner argv is shell-quoted text after `exec` in the wrapped script
rather than being the bare argv list), and made
_apptainer_runtime.py's test `_extract_inner_argv` unwrap helper
recursive since the D2 preflight wrapper's own exec line is itself this
new bash -lc wrapper (nested, not flat).
